single_app/requirements.txt
# requirements.txt

Flask==2.2.5
gunicorn
Werkzeug==3.0.6
requests==2.32.0
openai==1.59.7
docx2txt==0.8
python-docx==0.8.11
Markdown==3.3.4
bleach==6.1.0
azure-cosmos==4.3.0
msal==1.31.0
Flask-Session==0.8.0
azure-ai-documentintelligence==1.0.0b4
numpy==2.1.1
scikit-learn==1.5.2
SciPy==1.14.1
joblib==1.4.2
threadpoolctl==3.5.0
azure-search-documents==11.4.0
python-dotenv==0.19.1
azure-ai-formrecognizer==3.3.3
pyjwt==2.9.0
pandas==2.2.3
markdown2==2.5.3
azure-mgmt-cognitiveservices==13.6.0
azure-identity==1.17.1
azure-ai-contentsafety==1.0.0
Werkzeug>=2.2.0
azure-storage-blob==12.18.2
python-docx==0.8.11
docx2txt==0.8
pandas==2.2.3
openpyxl==3.1.2
python-pptx==0.6.21